Add GenericByteBuilder (#2969)
